compiled actions and functions to clean all imports and removed useless code
