+++ src/actions/index.js
@@ -1,0 +1,18 @@
+import { BOARD, setBoard } from './board'
+import { CLOSE_SHIP, closestShip } from './closeShip'
+import { GAME, startGame } from './gameState'
+import { FORM_GUESS, formGuess } from './formGuess'
+import { GUESS, guess } from './guess'
+
+export {
+  BOARD,
+  setBoard,
+  CLOSE_SHIP,
+  closestShip,
+  GAME,
+  startGame,
+  FORM_GUESS,
+  formGuess,
+  GUESS,
+  guess,
+}
