--- src/actions/ships.js
@@ -1,8 +1,0 @@
-export const BOTH_SHIPS = 'BOTH_SHIPS'
-
-export function bothShipPos(arrShips) {
-  return {
-    type: BOTH_SHIPS,
-    arrShips,
-  }
-}
