+++ src/components/GuessForm.jsx
@@ -1,21 +1,31 @@
-import React, { useState } from "react"
+/* eslint-disable react-hooks/exhaustive-deps */
+import React, { useEffect, useState } from "react"
 import { useDispatch, useSelector } from 'react-redux'
 import { FormLabel, VStack, Input } from "@chakra-ui/react";
 
-import { checkDist } from "../funcs/func";
-
-import {guess} from '../actions/guess'
-import {formGuess} from '../actions/oneGuess'
-import { closestShip } from '../actions/closeShip'
+import { shipCount, checkDist, getShipPositions } from "../funcs";
+import { formGuess, closestShip, startGame } from '../actions'
 
 function GuessForm() {
-  const ships = useSelector(state => state.shipPos)
+  const board = useSelector(state => state.board)
+  const guesses = useSelector(state => state.guesses)
 
   const dispatch = useDispatch()
   const [coords, setCoords] = useState({
     pos: ''
   })
+  
+  function checkIfGameCompleted() {
+    if ((shipCount(getShipPositions(board).length)) === 0) {
+      dispatch(startGame(false))
+    } 
+    return 'Game still going'
+  }
 
+  useEffect(() => {
+    checkIfGameCompleted()
+  },[guesses])
+  
   function changeHandler(e) {
     setCoords({
       ...coords,
@@ -26,11 +36,10 @@ function GuessForm() {
   function submitHandler(e) {
     e.preventDefault()
     let splitStr = coords.pos.split(',')
-    if(splitStr[0] >= 1 && splitStr[0] <= 8 && splitStr[1] >= 1 && splitStr[1] <= 8) {
+    if (splitStr[0] >= 1 && splitStr[0] <= 8 && splitStr[1] >= 1 && splitStr[1] <= 8) {
       let stringsToNum = {row: parseInt(splitStr[0]), col: parseInt(splitStr[1])}
-      dispatch(guess(stringsToNum))
       dispatch(formGuess(stringsToNum))
-      dispatch(closestShip(checkDist(ships, stringsToNum)))
+      dispatch(closestShip(checkDist(getShipPositions(board), stringsToNum)))
       setCoords({
         pos: ''
       })
@@ -38,6 +47,7 @@ function GuessForm() {
       alert('Needs to be in the format x,y and between 1 and 8!')
     }
   }
+
   return (
     <VStack className="form">
       <form onSubmit={submitHandler}>
