+++ src/funcs/getShipPositions.js
@@ -2,7 +2,7 @@ export const getShipPositions = (board) => {
   let shipArr = []
 
   board.map((tile) => {
-    if (tile.isShip === true) {
+    if (tile.isShip === true && tile.isVisible === false) {
       shipArr.push(tile)
     } else {
       return 'no ship'
