+++ src/funcs/index.js
@@ -1,0 +1,21 @@
+import { getShips } from './getShips'
+import { checkDist } from './checkDistance'
+import { addShips } from './createShips'
+import { getShipPositions } from './getShipPositions'
+import { shipTemp } from './getShipTemp'
+import { findShipPositions } from './makeShip'
+import { checkEdge } from './checkEdge'
+import { shipCount } from './shipCount'
+import { createBoard } from './genBoard'
+
+export {
+  getShips,
+  checkDist,
+  addShips,
+  getShipPositions,
+  shipTemp,
+  findShipPositions,
+  checkEdge,
+  shipCount,
+  createBoard,
+}
