+++ src/reducers/closeShip.js
@@ -1,8 +1,6 @@
 import { CLOSE_SHIP } from '../actions/closeShip'
 
-const nearestShip = ''
-
-export default function nearbyShip(state = nearestShip, action) {
+export default function nearbyShip(state = '', action) {
   switch (action.type) {
     case CLOSE_SHIP:
       return action.number
